+++ tests/testthat/test-bfsl.R
@@ -75,8 +75,8 @@ test_that("print method works", {
 Call:
 bfsl.default(x = pearson_york_data)
 
-Coefficients:\n           Estimate  Std. Error
-Intercept   5.47991   0.29497  \nSlope      -0.48053   0.05799  \n"
+Coefficients:\n             Estimate  Std. Error
+(Intercept)   5.47991   0.29497  \nSlope        -0.48053   0.05799  \n"
 , fixed=TRUE)
 })
 
@@ -102,9 +102,13 @@ test_that("summary method works", {
   expect_output(print(summary(fit)),
 "
 Call:
-bfsl.default(x = pearson_york_data)\n
+bfsl.default(x = pearson_york_data)
+
 Residuals:\n    Min.   1st Qu.    Median      Mean   3rd Qu.      Max.  \n-0.42396  -0.20650   0.14745   0.05573   0.34804   0.42009  \n
-Coefficients:\n           Estimate  Std. Error\nIntercept   5.47991   0.29497  \nSlope      -0.48053   0.05799  \n
+Coefficients:\n            Estimate Std. Error t value Pr(>|t|)    \n(Intercept)  5.47991    0.29497  18.578 7.27e-08 ***\nSlope       -0.48053    0.05799  -8.287 3.38e-05 ***
+---
+Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
+
 Goodness of fit: 1.483
 Chisq-statistic: 11.87 on 8 degrees of freedom
 Covariance of the slope and intercept: -0.01651